+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+<<<<<<< HEAD
+=======
+## [] - 2024-01-16
+- Issue 156: Updated FDO components to version 1.1.7
+
+## [] - 2023-11-13
+- Issue 152: Updated FDO components to version 1.1.6
+- Fixed package name typo in Fedora.
+- Changed docker-compose download to pull the minimum allowed version.
+
+>>>>>>> 06b87c9f04018125eb2d02cf7ff1f2913369e8b6
 ## [] - 2023-07-10
 - Issue 146: Integrated FDO into all-in-1.
 - Removed SDO Support.